Split content of interface.ts
