+++ src/ihttp_extension.ts
@@ -1,0 +1,4 @@
+export interface IHttpExtension {
+  readonly server: any;
+  readonly app: any;
+}
