+++ src/ihttp_router.ts
@@ -1,0 +1,6 @@
+export interface IHttpRouter {
+  readonly router: any;
+  readonly baseRoute: string;
+  initialize(): Promise<any> | any;
+  initializeRouter(): Promise<any> | any;
+}
