+++ src/index.ts
@@ -1,1 +1,5 @@
-export * from './interfaces';
+export * from './ihttp_client';
+export * from './ihttp_extension';
+export * from './ihttp_router';
+export * from './irequest_options';
+export * from './iresponse';
