+++ src/irequest_options.ts
@@ -1,0 +1,3 @@
+import * as popsicle from 'popsicle';
+
+export interface IRequestOptions extends popsicle.RequestOptions {}
