+++ src/iresponse.ts
@@ -1,0 +1,4 @@
+export interface IResponse<T> {
+  result: T | Array<T>;
+  status: number;
+}
